Moved V3 settings to the settings directory, removing the V3 one
