+++ src/settings/ExcludeRangeSettingV3.cpp
